remove multiside_uniforms, was used only by 'ff' faction
Overriding uniforms like this is not an effective way how to do it,
as it would involve a lot of manual work.

If something similar is needed in the future, create a more generic
automatic generation tool in sqf that would output the necessary
modification config structure for all uniforms in the game, also
automatically generating requiredAddons.

Note that this would make all soldiers which use that uniform have
an icon of this mod in Eden, even despite addonRootClass[].

Signed-off-by: freghar <freghar@dummy.tld>